apps/ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
requests
python-dotenv
google-generativeai